requirements.txt
pathlib
numpy
matplotlib
pandas
pyyaml
scipy